Fix for the time issues
When the server and the client work in different timezones the date sent
to the client is wrongly parsed since there is no TZ info in it. This
commit changes the way we store the hit date - from Date object to a
long value. This makes it easy to compare for the client.